2.9.9 Small improvements

+++ app/src/main/java/com/zilverline/domofence/GoogleApiBuilder.java
@@ -13,14 +13,12 @@ import android.support.annotation.NonNull;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.LocalBroadcastManager;
 import android.util.Log;
+import android.widget.EditText;
 import android.widget.Toast;
 
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.api.GoogleApiClient;
-import com.google.android.gms.common.api.ResultCallback;
-import com.google.android.gms.common.api.Status;
 import com.google.android.gms.location.Geofence;
-import com.google.android.gms.location.GeofenceStatusCodes;
 import com.google.android.gms.location.GeofencingRequest;
 import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -29,7 +27,7 @@ import com.google.android.gms.tasks.Task;
 import java.util.ArrayList;
 
 public class GoogleApiBuilder extends BroadcastReceiver implements
-        GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener, ResultCallback<Status> {
+        GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
 
     private static final String TAG = "GoogleApiBuilder";
     private static final String PACKAGENAME = "com.zilverline.domofence";
@@ -37,6 +35,7 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
     private GoogleApiClient mGoogleApiClient;
     private ArrayList<Geofence> mGeofenceList;
     private boolean mGeofencesAdded, start_after_boot;
+
     private PendingIntent mGeofencePendingIntent;
     private SharedPreferences mSharedPreferences;
     private String server_address, server_port, username, password, latitude, longitude, geofence_radius, idx_of_switch, protocol;
@@ -45,29 +44,33 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
 
     @Override
     public void onReceive(Context context, Intent intent) {
-        initialize(context);
-        Log.v(TAG, "onReceive after reboot");
+        if (Intent.ACTION_BOOT_COMPLETED.equals(intent.getAction())) {
+            Log.v(TAG, "onReceive after reboot");
+            initialize(context);
 
-        if (mGeofencesAdded) {
-            Log.v(TAG, "Geofence was added, re-adding");
+            if (mGeofencesAdded) {
+                Log.v(TAG, "Geofence was added, re-adding");
 
-            server_address = mSharedPreferences.getString(PACKAGENAME + ".server_address", "not_found");
-            server_port = mSharedPreferences.getString(PACKAGENAME + ".server_port", "not_found");
-            username = mSharedPreferences.getString(PACKAGENAME + ".username", "not_found");
-            password = mSharedPreferences.getString(PACKAGENAME + ".password", "not_found");
-            latitude = mSharedPreferences.getString(PACKAGENAME + ".latitude", "not_found");
-            longitude = mSharedPreferences.getString(PACKAGENAME + ".longitude", "not_found");
-            geofence_radius = mSharedPreferences.getString(PACKAGENAME + ".geofence_radius", "not_found");
-            idx_of_switch = mSharedPreferences.getString(PACKAGENAME + ".idx_of_switch", "not_found");
-            protocol = mSharedPreferences.getString(PACKAGENAME + ".protocol", "not_found");
+                server_address = mSharedPreferences.getString(PACKAGENAME + ".server_address", "not_found");
+                server_port = mSharedPreferences.getString(PACKAGENAME + ".server_port", "not_found");
+                username = mSharedPreferences.getString(PACKAGENAME + ".username", "not_found");
+                password = mSharedPreferences.getString(PACKAGENAME + ".password", "not_found");
+                latitude = mSharedPreferences.getString(PACKAGENAME + ".latitude", "not_found");
+                longitude = mSharedPreferences.getString(PACKAGENAME + ".longitude", "not_found");
+                geofence_radius = mSharedPreferences.getString(PACKAGENAME + ".geofence_radius", "not_found");
+                idx_of_switch = mSharedPreferences.getString(PACKAGENAME + ".idx_of_switch", "not_found");
+                protocol = mSharedPreferences.getString(PACKAGENAME + ".protocol", "not_found");
 
-            start_after_boot = true;
-            mGoogleApiClient.connect();
+                start_after_boot = true;
+                mGoogleApiClient.connect();
 
+            } else {
+                start_after_boot = false;
+                Log.v(TAG, "No data found, not initializing geofence");
+            }
         } else {
-            Log.v(TAG, "No data found, not initializing geofence");
+            Toast.makeText(baseContext, "Got unknown message: " + intent.getAction(), Toast.LENGTH_SHORT).show();
         }
-
     }
 
     protected void initialize(Context context) {
@@ -76,7 +79,6 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
 
         baseContext = context;
         mGeofenceList = new ArrayList<>();
-        mGeofencePendingIntent = null;
         mSharedPreferences = baseContext.getSharedPreferences(PACKAGENAME + ".SHARED_PREFERENCES_NAME",
                 Context.MODE_PRIVATE);
         mGeofencesAdded = mSharedPreferences.getBoolean(PACKAGENAME + ".GEOFENCES_ADDED_KEY", false);
@@ -85,11 +87,17 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
             throw new LocationServiceNoPermission();
         }
 
-        buildGoogleApiClient();
+        mGoogleApiClient = new GoogleApiClient.Builder(baseContext)
+                .addConnectionCallbacks(this)
+                .addOnConnectionFailedListener(this)
+                .addApi(LocationServices.API)
+                .build();
+
+        Log.v(TAG, "Build the Google API Client");
     }
 
     protected void addGeofence(String server_address, String server_port, String password, String username, String latitude, String longitude, String geofence_radius, String idx_of_switch, String protocol) {
-        Log.v(TAG, "Adding a geofence.  server_address: " + server_address +
+        Log.v(TAG, "Adding a geofence: " + mGeofencesAdded +".  server_address: " + server_address +
                 " server_port: " + server_port +
                 " password: " + password +
                 " username: " + username +
@@ -150,7 +158,8 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
     }
 
     private PendingIntent getGeofencePendingIntent() {
-        if (mGeofencePendingIntent != null) {
+
+        if (mGeofencePendingIntent != null){
             return mGeofencePendingIntent;
         }
         Intent intent = new Intent(baseContext, DomoFenceService.class);
@@ -163,8 +172,8 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
 
         Log.v(TAG, "Sending the Geofence Intent");
 
-
-        return PendingIntent.getService(baseContext, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
+        mGeofencePendingIntent = PendingIntent.getService(baseContext, 1977, intent, PendingIntent.FLAG_UPDATE_CURRENT);
+        return mGeofencePendingIntent;
     }
 
     public void populateGeofenceList() {
@@ -187,27 +196,28 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
                 .build());
     }
 
-    private synchronized void buildGoogleApiClient() {
-        mGoogleApiClient = new GoogleApiClient.Builder(baseContext)
-                .addConnectionCallbacks(this)
-                .addOnConnectionFailedListener(this)
-                .addApi(LocationServices.API)
-                .build();
-
-        Log.v(TAG, "Building the Google API Client");
-    }
-
-    protected Location getCurrentLocation() {
+    protected void getCurrentLocation(final EditText lat, final EditText lon) {
 
         if (mGoogleApiClient != null) {
             if (ActivityCompat.checkSelfPermission(baseContext, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(baseContext, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
                 throw new LocationServiceNoPermission();
             }
-            return LocationServices.getFusedLocationProviderClient(baseContext).getLastLocation().getResult();
+            LocationServices.getFusedLocationProviderClient(baseContext).getLastLocation().addOnCompleteListener(new OnCompleteListener<Location>() {
+                @Override
+                public void onComplete(@NonNull Task<Location> task) {
+                    if (task.isSuccessful() && task.getResult() != null) {
+                        lat.setText(String.valueOf(task.getResult().getLatitude()));
+                        lon.setText((String.valueOf(task.getResult().getLongitude())));
+                        Toast.makeText(baseContext, "Found and set current location!", Toast.LENGTH_SHORT).show();
+                    } else {
+                        Toast.makeText(baseContext, "Could not find current location!", Toast.LENGTH_SHORT).show();
+                    }
+
+                }
+            });
         } else {
-            Log.e(TAG, "Cant get currentlocation, GoogleApiCLient not ready");
+            Log.e(TAG, "Cant get currentlocation, GoogleApiClient not ready");
             Toast.makeText(baseContext, "Google client not (yet) ready, can't get current location.", Toast.LENGTH_SHORT).show();
-            return null;
         }
 
     }
@@ -225,7 +235,6 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
         if(start_after_boot){
             addGeofence(server_address, server_port, password, username, latitude, longitude, geofence_radius, idx_of_switch, protocol);
         }
-
     }
 
     @Override
@@ -237,55 +246,6 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
     @Override
     public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
         Log.i(TAG, "Connection failed: ConnectionResult.getErrorCode() = " + connectionResult.getErrorCode());
-    }
-
-    @Override
-    public void onResult(@NonNull Status status) {
-        if (status.isSuccess() && ! start_after_boot) {
-            mGeofencesAdded = !mGeofencesAdded;
-            SharedPreferences.Editor editor = mSharedPreferences.edit();
-            editor.putBoolean(PACKAGENAME + ".GEOFENCES_ADDED_KEY", mGeofencesAdded);
-            editor.putString(PACKAGENAME + ".server_address", server_address);
-            editor.putString(PACKAGENAME + ".server_port", server_port);
-            editor.putString(PACKAGENAME + ".username", username);
-            editor.putString(PACKAGENAME + ".password", password);
-            editor.putString(PACKAGENAME + ".longitude", longitude);
-            editor.putString(PACKAGENAME + ".latitude", latitude);
-            editor.putString(PACKAGENAME + ".geofence_radius", geofence_radius);
-            editor.putString(PACKAGENAME + ".idx_of_switch", idx_of_switch);
-            editor.putString(PACKAGENAME + ".protocol", protocol);
-
-            editor.apply();
-
-            if(mGeofencesAdded) {
-                sendStatus("Added");
-            } else {
-                sendStatus("Removed");
-            }
-
-            Toast.makeText(
-                    baseContext,
-                    mGeofencesAdded ? "The Geofence was added" : "All Geofences are removed",
-                    Toast.LENGTH_SHORT
-            ).show();
-        } else {
-            String errorMessage;
-
-            switch (status.getStatusCode()) {
-                case GeofenceStatusCodes.GEOFENCE_NOT_AVAILABLE:
-                    errorMessage = "Geofence service is not available now"; break;
-                case GeofenceStatusCodes.GEOFENCE_TOO_MANY_GEOFENCES:
-                    errorMessage = "Your app has registered too many geofences"; break;
-                case GeofenceStatusCodes.GEOFENCE_TOO_MANY_PENDING_INTENTS:
-                    errorMessage = "You have provided too many PendingIntents to the addGeofences() call"; break;
-                default:
-                    errorMessage = "Unknown error: the Geofence service is not available now. " + status.getStatusCode();
-            }
-            Toast.makeText(baseContext, errorMessage, Toast.LENGTH_SHORT).show();
-            Log.e(TAG, errorMessage);
-        }
-
-        start_after_boot = false;
     }
 
     private void sendStatus(String message) {
@@ -305,7 +265,7 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
     public class GeofenceCompletedListener implements OnCompleteListener {
         @Override
         public void onComplete(@NonNull Task task) {
-            if (task.isSuccessful() && ! start_after_boot) {
+            if (task.isSuccessful() & !start_after_boot) {
                 mGeofencesAdded = !mGeofencesAdded;
                 SharedPreferences.Editor editor = mSharedPreferences.edit();
                 editor.putBoolean(PACKAGENAME + ".GEOFENCES_ADDED_KEY", mGeofencesAdded);
@@ -320,23 +280,23 @@ public class GoogleApiBuilder extends BroadcastReceiver implements
                 editor.putString(PACKAGENAME + ".protocol", protocol);
 
                 editor.apply();
-
-                if(mGeofencesAdded) {
-                    sendStatus("Added");
-                } else {
-                    sendStatus("Removed");
-                }
-
-                Toast.makeText(
-                        baseContext,
-                        mGeofencesAdded ? "The Geofence was added" : "All Geofences are removed",
-                        Toast.LENGTH_SHORT
-                ).show();
-            } else {
+            } else if (!task.isSuccessful()){
                 String errorMessage = "Error while registering a geofence: " + task.getResult();
                 Toast.makeText(baseContext, errorMessage, Toast.LENGTH_SHORT).show();
                 Log.e(TAG, errorMessage);
             }
+
+            if(mGeofencesAdded) {
+                sendStatus("Added");
+            } else {
+                sendStatus("Removed");
+            }
+
+            Toast.makeText(
+                    baseContext,
+                    mGeofencesAdded ? "The Geofence was added" : "All Geofences are removed",
+                    Toast.LENGTH_SHORT
+            ).show();
 
             start_after_boot = false;
         }
